#19 Updates as per insights
