+++ app/Policies/PartPolicy.php
@@ -4,7 +4,6 @@ namespace App\Policies;
 
 use App\Models\Part;
 use App\Models\User;
-use Illuminate\Auth\Access\Response;
 
 class PartPolicy
 {
